+++ changelog.md
@@ -2,6 +2,7 @@
 - Made sure word boundaries were respected by morse code codec.
 - Fix session restore bug
 - Fix bug that could inadvertantly wipe document data if AES or RSA fails.
+- Fix issue with the browse buttons in settings for the encryption keys
 
 2026/07/25
 - Fixed issue when selecting piglatin from the toolbar.